fixed 'no such process' and now look at first 15 lines

+++ sh/daekwtr.php
@@ -58,8 +58,7 @@ class kwProcMutex {
 	$pid = trim(file_get_contents($this->pidfn));
 	if (!$pid || !is_numeric($pid) || !preg_match('/^\d+$/', $pid)) return;
 	$r = trim(shell_exec("kill -15 $pid 2>&1 "));
+	if (strpos($r, 'No such process') !== false) return;
 	if ($r) die('kill failed with ' . $r);
-	// if (strpos($r, 'No such process')) return;
-	// die('already running' . "\n");
     }
 }
